:construction: update celery functions
